Move "back" functions in /server
